Add support for DMaaP in SMO
SMO is planning on using Kafka bus as the message bus. DMaaP is nothing but a wrapper around the Kafka bus. Since there are legacy applications, mostly from ONAP that need support for reading from or writing to the message bus using DMaaP. This task will track the work related to that support.

Issue-Id: SMO-18
Signed-off-by: santanude <santanu.de@xoriant.com>
Change-Id: I31303440051fd5823ec2bb3199011d3920de6e13